refactor: personal-information 컴포넌트 분리
